+++ src/web/assets/cp/src/js/DraftEditor.js
@@ -1237,27 +1237,39 @@ Craft.DraftEditor = Garnish.Base.extend(
      * @returns {string}
      */
     swapDuplicatedElementIds: function (data) {
+      if (!Craft.fieldsWithoutContent.length) {
+        return data;
+      }
       const idsRE = Object.keys(this.duplicatedElements).join('|');
       if (idsRE === '') {
         return data;
       }
       const lb = encodeURIComponent('[');
       const rb = encodeURIComponent(']');
+      const namePrefixRE = `fields${lb}(?:${Craft.fieldsWithoutContent.join(
+        '|'
+      )})${rb}`;
       // Keep replacing field IDs until data stops changing
       while (true) {
         if (
           data ===
           (data = data
-            // &fields[...][X]
+            // &fields[handle]([...])?[X]
             .replace(
-              new RegExp(`(&fields${lb}[^=]+${rb}${lb})(${idsRE})(${rb})`, 'g'),
+              new RegExp(
+                `(&${namePrefixRE}(?:${lb}[^=]+${rb})?${lb})(${idsRE})(${rb})`,
+                'g'
+              ),
               (m, pre, id, post) => {
                 return pre + this.duplicatedElements[id] + post;
               }
             )
-            // &fields[...=X
+            // &fields[handle]([...)?=X
             .replace(
-              new RegExp(`&(fields${lb}[^=]+)=(${idsRE})\\b`, 'g'),
+              new RegExp(
+                `&(${namePrefixRE}(?:${lb}[^=]+)?)=(${idsRE})\\b`,
+                'g'
+              ),
               (m, name, id) => {
                 // Ignore param names that end in `[enabled]`, `[type]`, etc.
                 // (`[sortOrder]` should pass here, which could be set to a specific order index, but *not* `[sortOrder][]`!)